+++ lib/cli_displayed/dis_cli.py
@@ -1,4 +1,5 @@
-def display_order(order):
+# private
+def _display_order(order):
     if order > 0:
         if order >1:
             set = 1
@@ -9,17 +10,19 @@ def display_order(order):
         order = ""
     return order
 
-    
+def _massage_func(typo,massage,order):
+    return f"{_display_order(order)}[{typo}] {massage}"
+
+
+# public
 def display_typo(order,func,massage,optional_massage="",when=False):
     typo = func
     if typo:
-        print(massage_func("/",massage,order))
+        print(_massage_func("/",massage,order))
     else:
-        print(massage_func("x",massage,order))
+        print(_massage_func("x",massage,order))
     if optional_massage != "":
         if typo == when:
-            print(massage_func("*",optional_massage,order+1))
+            print(_massage_func("*",optional_massage,order+1))
 
 
-def massage_func(typo,massage,order):
-    return f"{display_order(order)}[{typo}] {massage}"
