+++ lib/function_network/func_network.py
@@ -1,18 +1,7 @@
-import os, sys, inspect, json
-currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
-parentdir = os.path.dirname(currentdir)
-sys.path.insert(0, parentdir)
-from file_management.createapikeyfile import SaveApiKey
-from file_management.file_management_lib import WorkEditor
-from file_management.configeditor import ConfigEditor
+import os, json
 import requests
-import json
-import sys
-import inspect
-currentdir = os.path.dirname(os.path.abspath(
-    inspect.getfile(inspect.currentframe())))
-parentdir = os.path.dirname(os.path.dirname(currentdir))
-sys.path.insert(0, parentdir)
+
+from lib.file_management import SaveApiKey, WorkEditor, ConfigEditor
 
 
 class Api:
