+++ src/main/pre_work.py
@@ -1,12 +1,5 @@
 from lib.file_management import WorkEditor
 import os
-import sys
-import inspect
-currentdir = os.path.dirname(os.path.abspath(
-    inspect.getfile(inspect.currentframe())))
-parentdir = os.path.dirname(os.path.dirname(currentdir))
-sys.path.insert(0, parentdir)
-
 
 class Work(WorkEditor):
     def __init__(self) -> None:
